refactoring p2 - more OOP-like experiment & runner in example
